+++ lua/telescope/finders/async_job_finder.lua
@@ -1,5 +1,5 @@
-local async_job = require "telescope._"
-local LinesPipe = require("telescope._").LinesPipe
+local async_job = require "telescope.async_job"
+local LinesPipe = async_job.LinesPipe
 
 local make_entry = require "telescope.make_entry"
 local log = require "telescope.log"
